Refactoring tools to be automatically loaded
